- Added light to center of screen - ADDED NEW LIBRARY: LIQUID FUN (for waves, box2d extension)

+++ desktop/src/com/dataproject/platforms/desktop/DesktopLauncher.java
@@ -8,7 +8,7 @@ public class DesktopLauncher {
 	public static void main (String[] arg)
 	{
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		config.height = (int)(500);
+		config.height = 500;
 		config.width = (int)(500*1.6180339);
